requirements.txt
aiohttp==3.11.14
Flask==2.2.2
jieba==0.42.1
lxml==4.2.4
openai==1.69.0
pandas==1.4.2
python-dotenv==1.1.0
Requests==2.32.3
scipy==1.7.3
lancedb==0.21.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.45
langchain-text-splitters==0.3.7
mysql-connector-python==9.2.0
numpy==2.0.2
pydantic==2.10.6
